Allow to change the allocator dynamically
