+++ lib/src/alloc.h
@@ -9,75 +9,27 @@ extern "C" {
 #include <stdbool.h>
 #include <stdio.h>
 
-#if defined(TREE_SITTER_ALLOCATION_TRACKING)
+struct allocator {
+  void *(*malloc)(size_t);
+  void *(*calloc)(size_t, size_t);
+  void *(*realloc)(void *, size_t);
+  void (*free)(void *);
+};
 
-void *ts_record_malloc(size_t);
-void *ts_record_calloc(size_t, size_t);
-void *ts_record_realloc(void *, size_t);
-void ts_record_free(void *);
-bool ts_toggle_allocation_recording(bool);
-
-#define ts_malloc  ts_record_malloc
-#define ts_calloc  ts_record_calloc
-#define ts_realloc ts_record_realloc
-#define ts_free    ts_record_free
-
-#else
+extern struct allocator *ts_allocator;
 
 // Allow clients to override allocation functions
-
 #ifndef ts_malloc
-#define ts_malloc  ts_malloc_default
+#define ts_malloc  ts_allocator->malloc
 #endif
 #ifndef ts_calloc
-#define ts_calloc  ts_calloc_default
+#define ts_calloc  ts_allocator->calloc
 #endif
 #ifndef ts_realloc
-#define ts_realloc ts_realloc_default
+#define ts_realloc ts_allocator->realloc
 #endif
 #ifndef ts_free
-#define ts_free    ts_free_default
-#endif
-
-#include <stdlib.h>
-
-static inline bool ts_toggle_allocation_recording(bool value) {
-  (void)value;
-  return false;
-}
-
-
-static inline void *ts_malloc_default(size_t size) {
-  void *result = malloc(size);
-  if (size > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to allocate %zu bytes", size);
-    exit(1);
-  }
-  return result;
-}
-
-static inline void *ts_calloc_default(size_t count, size_t size) {
-  void *result = calloc(count, size);
-  if (count > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to allocate %zu bytes", count * size);
-    exit(1);
-  }
-  return result;
-}
-
-static inline void *ts_realloc_default(void *buffer, size_t size) {
-  void *result = realloc(buffer, size);
-  if (size > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to reallocate %zu bytes", size);
-    exit(1);
-  }
-  return result;
-}
-
-static inline void ts_free_default(void *buffer) {
-  free(buffer);
-}
-
+#define ts_free    ts_allocator->free
 #endif
 
 #ifdef __cplusplus
